Created modules to separete codes because of linter error
