+++ src/index.js
@@ -1,5 +1,6 @@
 import './style.css';
-import { countDisplayedShows, countShows } from './modules/counterItem';
+import { countDisplayedShows, countShows } from './modules/counterItem.js';
+import createShowCard from './modules/display.js';
 
 const API = 'https://api.tvmaze.com/shows/1/episodes';
 const likesAPI = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/OpXkQhwbfD4wnLSWy6wV/likes';
@@ -11,7 +12,6 @@ const pageNum = document.querySelector('.page-numbers');
 const showsHeader = document.querySelector('.shows-header');
 const body = document.querySelector('body');
 
-
 let showsArray = [];
 
 const getData = async (API) => {
@@ -19,62 +19,15 @@ const getData = async (API) => {
   return result;
 };
 
-const likeShow = async (id, likesNumber, likesBtn) => {
-  await fetch(likesAPI, {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: id,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  }).then(() => {
-    likesNumber.innerHTML = parseInt(likesNumber.innerHTML, 10) + 1;
-    const i = likesBtn.querySelector('i');
-    i.classList.remove('fa-regular');
-    i.classList.add('fa-solid');
+const displayShows = (shows, pageNumber, heading) => {
+  shows.slice(pageNumber * 10 - 10, pageNumber * 10).forEach((show) => {
+    const div = createShowCard(show);
+    displayEpisode.append(div);
   });
+  const displayedShowsObj = countDisplayedShows(displayEpisode);
+  const showsCount = countShows(showsArray);
+  heading.innerHTML = `Episodes: (${displayedShowsObj.firstId}, ${displayedShowsObj.lastId}) of ${showsCount}`;
 };
-
-function createShowCard(episode) {
-  const div = document.createElement('div');
-  div.classList.add('card');
-  div.setAttribute('id', episode.id);
-  div.innerHTML = `
-  <h2 class="card-title">${episode.name}</h2>
-  <img class="card-img-top img" src=${episode.image.medium} alt="">
-  <div class="card-body">       
-     <p class="card-text">${episode.summary}</p>
-     <button type="button" class="btn btn-link btn-comment">Comments</i></button>
-     <div class="likes">
-     <button type="button"
-      class="btn btn-link like-count"><i class="fa-regular fa-heart like-icon"></i></button>
-     <span class="likeCount">0</span>
-     </div>
-     </div>
-  `;
-  const likesNumber = div.querySelector('.likeCount');
-  if ('likes' in episode) {
-    likesNumber.innerHTML = episode.likes;
-  }
-
-  const likeBtn = div.querySelector('.like-count');
-  likeBtn.addEventListener('click', () => {
-    likeShow(episode.id, likesNumber, likeBtn);
-  });
-
-  return div;
-}
-
-  const displayShows = (shows, pageNumber, heading) => {
-    shows.slice(pageNumber * 10 - 10, pageNumber * 10).forEach((show) => {
-      const div = createShowCard(show);
-      displayEpisode.append(div);
-    });
-    const displayedShowsObj = countDisplayedShows(displayEpisode);
-    const showsCount = countShows(showsArray);
-    heading.innerHTML = `Episodes: (${displayedShowsObj.firstId}, ${displayedShowsObj.lastId}) of ${showsCount}`;
-  };
 
 function loadNext(pageNumber, shows, showsHeader) {
   const nextPage = pageNumber + 1;
@@ -108,8 +61,7 @@ next.addEventListener('click', () => {
   loadNext(parseInt(pageNum.innerHTML, 10), showsArray, showsHeader, pageNum, displayEpisode, body, likesAPI);
 });
 
-
-  getData(API)
+getData(API)
   .then((response1) => response1.json())
   .then((shows) => {
     showsArray = shows;
