--- src/modules/fetch.js
@@ -1,4 +1,0 @@
-export const getShowsData = async (url) => {
-    const result = await fetch(url);
-    return result;
-  };
